rebase to master

+++ nova/network/api.py
@@ -18,7 +18,9 @@
 
 import functools
 
-from oslo.config import cfg
+from oslo_config import cfg
+from oslo_log import log as logging
+from oslo_utils import strutils
 
 from nova import exception
 from nova.i18n import _LI
@@ -28,7 +30,6 @@ from nova.network import model as network_model
 from nova.network import rpcapi as network_rpcapi
 from nova import objects
 from nova.objects import base as obj_base
-from nova.openstack.common import log as logging
 from nova import policy
 from nova import utils
 
@@ -43,7 +44,8 @@ def wrap_check_policy(func):
     @functools.wraps(func)
     def wrapped(self, context, *args, **kwargs):
         action = func.__name__
-        check_policy(context, action)
+        if not self.skip_policy_check:
+            check_policy(context, action)
         return func(self, context, *args, **kwargs)
 
     return wrapped
@@ -94,7 +96,7 @@ class API(base_api.NetworkAPI):
 
     @wrap_check_policy
     def get(self, context, network_uuid):
-        return objects.Network.get_by_uuid(context.elevated(), network_uuid)
+        return objects.Network.get_by_uuid(context, network_uuid)
 
     @wrap_check_policy
     def create(self, context, **kwargs):
@@ -102,6 +104,9 @@ class API(base_api.NetworkAPI):
 
     @wrap_check_policy
     def delete(self, context, network_uuid):
+        network = self.get(context, network_uuid)
+        if network.project_id is not None:
+            raise exception.NetworkInUse(network_id=network_uuid)
         return self.network_rpcapi.delete_network(context, network_uuid, None)
 
     @wrap_check_policy
@@ -120,7 +125,7 @@ class API(base_api.NetworkAPI):
 
     @wrap_check_policy
     def get_floating_ip(self, context, id):
-        if not utils.is_int_like(id):
+        if not strutils.is_int_like(id):
             raise exception.InvalidID(id=id)
         return objects.FloatingIP.get_by_id(context, id)
 
@@ -313,9 +318,9 @@ class API(base_api.NetworkAPI):
     def add_fixed_ip_to_instance(self, context, instance, network_id):
         """Adds a fixed ip to instance from specified network."""
         flavor = instance.get_flavor()
-        args = {'instance_id': instance['uuid'],
+        args = {'instance_id': instance.uuid,
                 'rxtx_factor': flavor['rxtx_factor'],
-                'host': instance['host'],
+                'host': instance.host,
                 'network_id': network_id}
         nw_info = self.network_rpcapi.add_fixed_ip_to_instance(
             context, **args)
@@ -327,9 +332,9 @@ class API(base_api.NetworkAPI):
         """Removes a fixed ip from instance from specified network."""
 
         flavor = instance.get_flavor()
-        args = {'instance_id': instance['uuid'],
+        args = {'instance_id': instance.uuid,
                 'rxtx_factor': flavor['rxtx_factor'],
-                'host': instance['host'],
+                'host': instance.host,
                 'address': address}
         nw_info = self.network_rpcapi.remove_fixed_ip_from_instance(
             context, **args)
@@ -376,10 +381,10 @@ class API(base_api.NetworkAPI):
     def _get_instance_nw_info(self, context, instance):
         """Returns all network info related to an instance."""
         flavor = instance.get_flavor()
-        args = {'instance_id': instance['uuid'],
+        args = {'instance_id': instance.uuid,
                 'rxtx_factor': flavor['rxtx_factor'],
-                'host': instance['host'],
-                'project_id': instance['project_id']}
+                'host': instance.host,
+                'project_id': instance.project_id}
         nw_info = self.network_rpcapi.get_instance_nw_info(context, **args)
 
         return network_model.NetworkInfo.hydrate(nw_info)
@@ -412,14 +417,6 @@ class API(base_api.NetworkAPI):
         """
         # This is NOOP for Nova network since it doesn't support SR-IOV.
         pass
-
-    @wrap_check_policy
-    def get_instance_uuids_by_ip_filter(self, context, filters):
-        """Returns a list of dicts in the form of
-        {'instance_uuid': uuid, 'ip': ip} that matched the ip_filter
-        """
-        return self.network_rpcapi.get_instance_uuids_by_ip_filter(context,
-                                                                   filters)
 
     @wrap_check_policy
     def get_dns_domains(self, context):
@@ -486,11 +483,11 @@ class API(base_api.NetworkAPI):
         """Setup or teardown the network structures on hosts related to
            instance.
         """
-        host = host or instance['host']
+        host = host or instance.host
         # NOTE(tr3buchet): host is passed in cases where we need to setup
         # or teardown the networks on a host which has been migrated to/from
-        # and instance['host'] is not yet or is no longer equal to
-        args = {'instance_id': instance['id'],
+        # and instance.host is not yet or is no longer equal to
+        args = {'instance_id': instance.id,
                 'host': host,
                 'teardown': teardown}
 
@@ -499,7 +496,7 @@ class API(base_api.NetworkAPI):
     def _get_multi_addresses(self, context, instance):
         try:
             fixed_ips = objects.FixedIPList.get_by_instance_uuid(
-                context, instance['uuid'])
+                context, instance.uuid)
         except exception.FixedIpNotFoundForInstance:
             return False, []
         addresses = []
@@ -513,9 +510,9 @@ class API(base_api.NetworkAPI):
         """Start to migrate the network of an instance."""
         flavor = instance.get_flavor()
         args = dict(
-            instance_uuid=instance['uuid'],
+            instance_uuid=instance.uuid,
             rxtx_factor=flavor['rxtx_factor'],
-            project_id=instance['project_id'],
+            project_id=instance.project_id,
             source_compute=migration['source_compute'],
             dest_compute=migration['dest_compute'],
             floating_addresses=None,
@@ -533,9 +530,9 @@ class API(base_api.NetworkAPI):
         """Finish migrating the network of an instance."""
         flavor = instance.get_flavor()
         args = dict(
-            instance_uuid=instance['uuid'],
+            instance_uuid=instance.uuid,
             rxtx_factor=flavor['rxtx_factor'],
-            project_id=instance['project_id'],
+            project_id=instance.project_id,
             source_compute=migration['source_compute'],
             dest_compute=migration['dest_compute'],
             floating_addresses=None,
@@ -547,3 +544,15 @@ class API(base_api.NetworkAPI):
             args['host'] = migration['dest_compute']
 
         self.network_rpcapi.migrate_instance_finish(context, **args)
+
+    def setup_instance_network_on_host(self, context, instance, host):
+        """Setup network for specified instance on host."""
+        self.migrate_instance_finish(context, instance,
+                                     {'source_compute': None,
+                                      'dest_compute': host})
+
+    def cleanup_instance_network_on_host(self, context, instance, host):
+        """Cleanup network for specified instance on host."""
+        self.migrate_instance_start(context, instance,
+                                    {'source_compute': host,
+                                     'dest_compute': None})
